% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check if character vector is a number
Regex "looks for decimal numbers, with optional HYPHEN MINUS and optional
FULL STOP plus zero or more decimal numbers following."

From https://stackoverflow.com/a/4247184/13945974}
\usage{
check_numeric(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A boolean
}
\description{
Check if character vector is a number
Regex "looks for decimal numbers, with optional HYPHEN MINUS and optional
FULL STOP plus zero or more decimal numbers following."

From https://stackoverflow.com/a/4247184/13945974
}
\author{
Shir Dekel
}
