% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_other.R
\name{check_other}
\alias{check_other}
\title{Conditional "other" text input display}
\usage{
check_other()
}
\value{
A string vector
}
\description{
Conditional "other" text input display
}
\author{
Shir Dekel
}
