% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_combined.R
\name{clean_data_combined}
\alias{clean_data_combined}
\title{Clean combined data}
\usage{
clean_data_combined(data_combined)
}
\arguments{
\item{data_combined}{Combined data}
}
\value{
A dataframe
}
\description{
Clean combined data
}
\author{
Shir Dekel
}
