% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_other.R
\name{clean_data_other}
\alias{clean_data_other}
\title{Clean other data}
\usage{
clean_data_other(data_raw_prep, main_stage)
}
\arguments{
\item{data_raw_prep}{Raw data}

\item{main_stage}{Main stage name}
}
\value{
A daydreamer
}
\description{
Clean other data
}
\author{
Shir Dekel
}
