% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_portfolio_binary.R
\name{clean_data_portfolio_binary}
\alias{clean_data_portfolio_binary}
\title{clean portfolio binary data}
\usage{
clean_data_portfolio_binary(data_raw_prep)
}
\arguments{
\item{data_raw_prep}{Raw data}
}
\value{
A dataframe
}
\description{
clean portfolio binary data
}
\author{
Shir Dekel
}
