% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_fromJSON.R
\name{column_fromJSON}
\alias{column_fromJSON}
\title{Convert string to JSON by column}
\usage{
column_fromJSON(data, columns)
}
\arguments{
\item{data}{Data}

\item{columns}{Columns}
}
\value{
A dataframe
}
\description{
Need to convert columns such as stage from JSON. Making sure it comes out
normal from jaysire proved to be difficult because it unboxes also other
elements.
}
\author{
Shir Dekel
}
