% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aggregated_values.R
\name{get_aggregated_values}
\alias{get_aggregated_values}
\title{Get aggregated values}
\usage{
get_aggregated_values(
  outcome_positive,
  prob_positive,
  outcome_dif,
  sort = TRUE
)
}
\arguments{
\item{outcome_positive}{Positive outcome}

\item{prob_positive}{Positive probability}

\item{outcome_dif}{Outcome difference}

\item{sort}{Whether to sort outcomes}
}
\value{
A list of numeric values
}
\description{
Get aggregated values
}
\author{
Shir Dekel
}
