% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_business_information.R
\name{get_business_information}
\alias{get_business_information}
\title{Get business information}
\usage{
get_business_information()
}
\value{
A "trial" object
}
\description{
Get business information
}
\author{
Shir Dekel
}
