% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consent_html.R
\name{get_consent_html}
\alias{get_consent_html}
\title{Generate consent HTML file

Template from jsPsych html-external example}
\usage{
get_consent_html(experiment_resources)
}
\arguments{
\item{experiment_resources}{Experiment resources path}
}
\description{
Generate consent HTML file

Template from jsPsych html-external example
}
\author{
Shir Dekel
}
