% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{Get demographic information}
\usage{
get_demographics()
}
\value{
A "trial" object
}
\description{
Get demographic information
}
\author{
Shir Dekel
}
