% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Get distribution

For aggregation gambles}
\usage{
get_distribution(outcome, probability, unit = "$")
}
\arguments{
\item{outcome}{Outcome values}

\item{probability}{Probability values}

\item{unit}{Unit for y-axis label}
}
\value{
ggplot object
}
\description{
Get distribution

For aggregation gambles
}
\author{
Shir Dekel
}
