% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment.R
\name{get_experiment}
\alias{get_experiment}
\title{Generate experiment files}
\usage{
get_experiment(
  gambles,
  experiment_directory,
  experiment_resources,
  main,
  post_experiment,
  columns,
  condition_allocation,
  ethics = TRUE,
  zip = TRUE,
  on_finish = save_psychserver()
)
}
\arguments{
\item{gambles}{Gamble values}

\item{experiment_directory}{Experiment directory}

\item{experiment_resources}{Experiment resources file path}

\item{main}{Main experiment objects}

\item{post_experiment}{Post experiment}

\item{columns}{Columns}

\item{condition_allocation}{Condition allocation}

\item{ethics}{Whether ethics pages are included (used for testing).}

\item{zip}{Whether to zip the experiment files}

\item{on_finish}{\code{on_finish} code}
}
\value{
Invisibly returns NULL
}
\description{
Generate experiment files
}
\author{
Shir Dekel
}
