% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_path_materials.R
\name{get_file_path_materials}
\alias{get_file_path_materials}
\title{Get materials file paths}
\usage{
get_file_path_materials(materials_directory, file_name_materials)
}
\arguments{
\item{materials_directory}{Materials directory}

\item{file_name_materials}{File name of the materials}
}
\value{
A filepath string
}
\description{
Get materials file paths
}
\author{
Shir Dekel
}
