% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gambles.R
\name{get_gambles}
\alias{get_gambles}
\title{Get gambles

For aggregation experiment generation}
\usage{
get_gambles(restricted_values, gamble_n)
}
\arguments{
\item{restricted_values}{Restricted values}

\item{gamble_n}{Number of gambles}
}
\value{
A list of numeric values
}
\description{
Get gambles

For aggregation experiment generation
}
\author{
Shir Dekel
}
