% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indefinite.R
\name{get_indefinite}
\alias{get_indefinite}
\title{Generate relevant indefinite given string}
\usage{
get_indefinite(str)
}
\arguments{
\item{str}{A string}
}
\value{
A string
}
\description{
Generate relevant indefinite given string
}
