% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loss_prob.R
\name{get_loss_prob}
\alias{get_loss_prob}
\title{Get loss probability}
\usage{
get_loss_prob(outcome_aggregated, prob_aggregated)
}
\arguments{
\item{outcome_aggregated}{Aggregated outcome}

\item{prob_aggregated}{Aggregated probability}
}
\value{
A numeric value
}
\description{
Get loss probability
}
\author{
Shir Dekel
}
