% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_min_difference.R
\name{get_max_min_difference}
\alias{get_max_min_difference}
\title{Get max/min NPV difference}
\usage{
get_max_min_difference(data, npv_amount, ...)
}
\arguments{
\item{data}{Data}

\item{npv_amount}{NPV amount}

\item{...}{Grouping variables}
}
\value{
A dataframe
}
\description{
Get max/min NPV difference
}
\author{
Shir Dekel
}
