% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pre_experiment.R
\name{get_pre_experiment}
\alias{get_pre_experiment}
\title{Get pre experiment trials}
\usage{
get_pre_experiment(ethics, experiment_resources)
}
\arguments{
\item{ethics}{Whether to include ethics (removed for testing).}

\item{experiment_resources}{Experiment resources file path}
}
\value{
A "timeline" object
}
\description{
Get pre experiment trials
}
\author{
Shir Dekel
}
