% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_positive_seq.R
\name{get_prob_positive_seq}
\alias{get_prob_positive_seq}
\title{Generate positive probability sequence

For aggregation experiment generation}
\usage{
get_prob_positive_seq()
}
\value{
A numeric vector
}
\description{
Generate positive probability sequence

For aggregation experiment generation
}
\author{
Shir Dekel
}
