% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_input.R
\name{get_project_input}
\alias{get_project_input}
\title{Get long project inputs

To name questions

For aggregation experiment generation}
\usage{
get_project_input(type, detail = NULL, gambles)
}
\arguments{
\item{type}{Project type}

\item{detail}{Project detail}

\item{gambles}{Gamble values}
}
\value{
A string
}
\description{
Get long project inputs

To name questions

For aggregation experiment generation
}
\author{
Shir Dekel
}
