% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restriction.R
\name{get_restriction}
\alias{get_restriction}
\title{Get restriction logicals

For aggregation experiment generation}
\usage{
get_restriction(restriction_values, gain_loss_ratio_restriction)
}
\arguments{
\item{restriction_values}{Restriction values}

\item{gain_loss_ratio_restriction}{Gain/loss ratio restriction}
}
\value{
A boolean
}
\description{
Get restriction logicals

For aggregation experiment generation
}
\author{
Shir Dekel
}
