% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restriction_values.R
\name{get_restriction_values}
\alias{get_restriction_values}
\title{Get restriction values (EV and gain/loss)

For aggregation experiment generation}
\usage{
get_restriction_values(prob_positive_sample, outcome_positive_sample)
}
\arguments{
\item{prob_positive_sample}{Positive probability}

\item{outcome_positive_sample}{Positive outcome}
}
\value{
A list of numeric values
}
\description{
Get restriction values (EV and gain/loss)

For aggregation experiment generation
}
\author{
Shir Dekel
}
