% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_number.R
\name{get_survey_number}
\alias{get_survey_number}
\title{Get number-type survey question}
\usage{
get_survey_number(
  label_text,
  name,
  min = 0,
  max = 100,
  width = 70,
  prefix = "",
  suffix = "",
  class = NULL
)
}
\arguments{
\item{label_text}{Label text}

\item{name}{Name}

\item{min}{Minimum value}

\item{max}{Minimum value}

\item{width}{Width}

\item{prefix}{Prefix}

\item{suffix}{Suffix}

\item{class}{Class}
}
\value{
A string marked as HTML
}
\description{
Get number-type survey question
}
