% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_select.R
\name{get_survey_select}
\alias{get_survey_select}
\title{Get drop-down survey question}
\usage{
get_survey_select(name_select, name_other, option, label_select)
}
\arguments{
\item{name_select}{Select name}

\item{name_other}{Other name}

\item{option}{Option}

\item{label_select}{Select label}
}
\value{
A string marked as HTML
}
\description{
Get drop-down survey question
}
\author{
Shir Dekel
}
