% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_portfolio_binary.R
\name{get_trial_portfolio_binary}
\alias{get_trial_portfolio_binary}
\title{Get portfolio trial - binary}
\usage{
get_trial_portfolio_binary(preamble = "", distribution = "absent")
}
\arguments{
\item{preamble}{Preamble}

\item{distribution}{Distribution}
}
\value{
A "trial" object
}
\description{
Get portfolio trial - binary
}
\author{
Shir Dekel
}
