% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_portfolio_number.R
\name{get_trial_portfolio_number}
\alias{get_trial_portfolio_number}
\title{Get portfolio trial - number}
\usage{
get_trial_portfolio_number(
  preamble = "",
  distribution = "absent",
  project_number = 10
)
}
\arguments{
\item{preamble}{Preamble}

\item{distribution}{Distribution}

\item{project_number}{Project number}
}
\value{
A "trial" object
}
\description{
Get portfolio trial - number
}
\author{
Shir Dekel
}
