% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_project_number.R
\name{get_trial_project_number}
\alias{get_trial_project_number}
\title{Project number follow up question}
\usage{
get_trial_project_number(max = 20)
}
\arguments{
\item{max}{Maximum value}
}
\value{
A "trial" object
}
\description{
Project number follow up question
}
\author{
Shir Dekel
}
