% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_separate.R
\name{get_trial_separate}
\alias{get_trial_separate}
\title{Get separate condition timeline

For aggregation experiment generation}
\usage{
get_trial_separate(
  preamble = "<p>Indicate below whether you would invest in the following:</p>",
  project_description,
  form_options = c("Yes", "No"),
  project_input,
  distribution,
  randomize_order
)
}
\arguments{
\item{preamble}{Preamble}

\item{project_description}{Project description}

\item{form_options}{Form options}

\item{project_input}{Project input}

\item{distribution}{Distribution}

\item{randomize_order}{Whether to randomize order}
}
\value{
An object of class "timeline"
}
\description{
Get separate condition timeline

For aggregation experiment generation
}
\author{
Shir Dekel
}
