% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_welcome.R
\name{get_welcome}
\alias{get_welcome}
\title{Get welcome page

For experiment generation}
\usage{
get_welcome()
}
\value{
A "trial" object
}
\description{
Get welcome page

For experiment generation
}
\author{
Shir Dekel
}
