% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_prolific_raw_data.R
\name{import_from_prolific_raw_data}
\alias{import_from_prolific_raw_data}
\title{Import from Prolific raw data}
\usage{
import_from_prolific_raw_data(
  prolific_raw_data,
  experiment_number,
  thesis_project
)
}
\arguments{
\item{prolific_raw_data}{File path of Prolific raw data}

\item{experiment_number}{Experiment number}

\item{thesis_project}{Thesis project}
}
\value{
A dataframe
}
\description{
Import from Prolific raw data
}
\author{
Shir Dekel
}
