% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choice.R
\name{plot_choice}
\alias{plot_choice}
\title{Plot choice data}
\usage{
plot_choice(data_effects, dv, dv_label)
}
\arguments{
\item{data_effects}{Data}

\item{dv}{DV}

\item{dv_label}{DV label}
}
\value{
A combined ggplot object
}
\description{
Plot choice data
}
\author{
Shir Dekel
}
