% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choice_trials.R
\name{plot_choice_trials}
\alias{plot_choice_trials}
\title{Get plot for choice x trial with optional grouping}
\usage{
plot_choice_trials(data, grouping = NULL)
}
\arguments{
\item{data}{Data}

\item{grouping}{Grouping variable}
}
\value{
A ggplot object
}
\description{
Get plot for choice x trial with optional grouping
}
\author{
Shir Dekel
}
