% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamble_values.R
\name{plot_gamble_values}
\alias{plot_gamble_values}
\title{Plot gamble values}
\usage{
plot_gamble_values(data)
}
\arguments{
\item{data}{Data}
}
\value{
A combined ggplot object
}
\description{
Plot gamble values
}
\author{
Shir Dekel
}
