% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gambles.R
\name{plot_gambles}
\alias{plot_gambles}
\title{Plot gambles

And save distribution figure}
\usage{
plot_gambles(gambles, experiment_resources, experiment_number)
}
\arguments{
\item{gambles}{Gamble values}

\item{experiment_resources}{Experiment resources file path}

\item{experiment_number}{Experiment number}
}
\value{
ggplot object
}
\description{
Plot gambles

And save distribution figure
}
\author{
Shir Dekel
}
