% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_project_number.R
\name{plot_project_number}
\alias{plot_project_number}
\title{Plot project number data}
\usage{
plot_project_number(data, condition = NULL)
}
\arguments{
\item{data}{Data}

\item{condition}{Condition name}
}
\value{
A plot
}
\description{
Plot project number data
}
\author{
Shir Dekel
}
