% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates

Group by Prolific ID, rank each ID's completion time, and only take the
first one.}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{Data}
}
\value{
A data frame
}
\description{
Remove duplicates

Group by Prolific ID, rank each ID's completion time, and only take the
first one.
}
\author{
Shir Dekel
}
