% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_psychserver.R
\name{save_psychserver}
\alias{save_psychserver}
\title{Return a javascript function to save data via a script on the psych webserver}
\usage{
save_psychserver()
}
\value{
An object of class "json"
}
\description{
Return a javascript function to save data via a script on the psych webserver
}
\author{
Shir Dekel
}
