% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shirthesis-package.R
\docType{package}
\name{shirthesis-package}
\alias{shirthesis}
\alias{shirthesis-package}
\title{shirthesis: Thesis functions}
\description{
Contains functions relevant to the thesis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shirdekel/shirthesis}
  \item Report bugs at \url{https://github.com/shirdekel/shirthesis/issues}
}

}
\author{
\strong{Maintainer}: Shir Dekel \email{shirdekel@yahoo.com.au} (\href{https://orcid.org/0000-0003-1773-2446}{ORCID})

}
\keyword{internal}
