% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_close.R
\name{verify_close}
\alias{verify_close}
\title{Insert verify close JS function

From Bruce Burns 23/07/20}
\usage{
verify_close()
}
\value{
An object of class "json"
}
\description{
Insert verify close JS function

From Bruce Burns 23/07/20
}
\author{
Shir Dekel
}
