% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shirthesis-package.R
\docType{package}
\name{shirthesis-package}
\alias{shirthesis}
\alias{shirthesis-package}
\title{shirthesis: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shir Dekel \email{shirdekel@yahoo.com.au} (\href{https://orcid.org/0000-0003-1773-2446}{ORCID})

}
\keyword{internal}
